move common logics to core package
