+++ packages/protoc-gen-nexus/src/dslgen/printers/util.ts
@@ -11,7 +11,6 @@ import {
   PrinterOptions,
   SquashedOneofUnionType,
 } from "@proto-graphql/codegen-core";
-import { ProtoField, ProtoMessage, ProtoScalarType } from "@proto-graphql/proto-descriptors";
 import { code, Code, def, imp, Import } from "ts-poet";
 
 export type NexusPrinterOptions = Extract<PrinterOptions, { dsl: "nexus" }>;
@@ -30,29 +29,6 @@ export function nexusTypeDef(
   return code`${def(type.typeName)}`;
 }
 
-const longScalarPrimitiveTypes: ReadonlySet<ProtoScalarType> = new Set([
-  "int64",
-  "uint64",
-  "fixed64",
-  "sfixed64",
-  "sint64",
-]);
-const longScalarWrapperTypes: ReadonlySet<string> = new Set([
-  "google.protobuf.Int64Value",
-  "google.protobuf.UInt64Value",
-]);
-
-export function isProtobufLong(proto: ProtoField): boolean {
-  switch (proto.type.kind) {
-    case "Scalar":
-      return longScalarPrimitiveTypes.has(proto.type.type);
-    case "Message":
-      return longScalarWrapperTypes.has(proto.type.fullName.toString());
-    default:
-      return false;
-  }
-}
-
 export function fieldType(
   field:
     | ObjectField<ObjectType | EnumType | InterfaceType | SquashedOneofUnionType>
@@ -63,8 +39,4 @@ export function fieldType(
   const importPath = generatedGraphQLTypeImportPath(field, opts);
   if (importPath) return code`${imp(`${field.type.typeName}@${importPath}`)}`;
   return code`${field.type.typeName}`;
-}
-
-export function isWellKnownType(proto: ProtoField["type"]): proto is ProtoMessage {
-  return proto.kind === "Message" && proto.file.name.startsWith("google/protobuf/");
 }
